heartguard_v3_2 (2)/heartguard_v3_2/heartguard_new/requirements.txt
# HeartGuard AI — Dependencies
# =============================

# Web Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.3

# ML & Data Science
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
joblib==1.4.2
matplotlib==3.10.1
seaborn==0.13.2

# PDF Generation
reportlab==4.4.0

# QR Code
qrcode==8.0
Pillow==11.2.1

# Notifications
twilio==9.6.0
python-dotenv==1.1.0

# Google OAuth
google-auth==2.38.0
requests==2.32.3
